+++ tests/integration/components/search-bar/component-test.ts
@@ -5,17 +5,13 @@ import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import { module, test } from 'qunit';
 
-module('Integration | Component | discover-research', hooks => {
+module('Integration | Component | search-bar', hooks => {
     setupRenderingTest(hooks);
     setupMirage(hooks);
 
     test('it renders', async function(assert) {
         this.set('search', {});
-        await render(hbs`<DiscoverResearch @onSearch={{this.search}} />`);
-        assert.dom('[data-test-discover-heading]').hasText('Discover public research');
-        assert.dom('[data-test-discover-subheading]')
-            .hasText('Discover projects, data, materials, and collaborators on OSF that might be helpful ' +
-            'to your own research.');
+        await render(hbs`<SearchBar @onSearch={{this.search}} />`);
         assert.dom('[data-test-search-bar]').exists();
         await a11yAudit(this.element);
         assert.ok(true, 'No a11y errors on page');
@@ -25,11 +21,11 @@ module('Integration | Component | discover-research', hooks => {
         assert.expect(2);
 
         this.set('search', () => { assert.ok(true); });
-        await render(hbs`<DiscoverResearch @onSearch={{action this.search}} />`);
+        await render(hbs`<SearchBar @onSearch={{action this.search}} />`);
 
-        await fillIn('[data-test-discover-input]', 'search string');
+        await fillIn('[data-test-search-input]', 'search string');
         settled().then(() => {
-            triggerKeyEvent('[data-test-discover-input]', 'keyup', 'Enter');
+            triggerKeyEvent('[data-test-search-input]', 'keyup', 'Enter');
         });
         await a11yAudit(this.element);
         assert.ok(true, 'No a11y errors on page');